+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+- new commands
+  - `expand`: expands the current item in the tree
+  - `expandcategories <scope>`: expand categories with articles in scope, i.e., `all`, `unread` or `marked` articles
+  - `collapse`: collapses the current item in the tree
+  - `collapseall`: collapses all items in the tree
+
 # 0.7.9 - 2026-01-22
 
 - removed share keybindings (`S m`, `S t`, etc.). instead `S` opens the command line with the configured share targets (press TAB to cycle through them)